pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyomslpwan"
version = "0.1.0"
authors = [
  { name="Maj Fontana Korošec", email="maj.fontana@gmail.com" },
]
description = "Bitstream generator and parser for OMS LPWAN physical layer"
#readme = "README.md"
requires-python = ">=3.8"
#classifiers = [
#    "Programming Language :: Python :: 3",
#    "License :: OSI Approved :: MIT License",
#    "Operating System :: OS Independent",
#]
dependencies = ["matplotlib",
                "numpy",
                "bitstring",
                "pyzmq",
                "numba",
                "streamz",
                "RsWaveform"]

[project.urls]
#Homepage = "https://github.com/pypa/sampleproject"
#Issues = "https://github.com/pypa/sampleproject/issues"